Changed Copyright content & Created Patient Reviews and Locations Page
