+++ src/App.js
@@ -1,6 +1,8 @@
 import Homepage from "./Components/homepage";
 import Navbar from "./Components/Navbar";
 import Footer from "./Components/footer";
+import Locations from "./Components/Locations";
+import PatientReviews from "./Components/PatientReviews";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
@@ -13,6 +15,8 @@ function App() {
     <div>
       <Navbar />
       <Homepage />
+      <PatientReviews />
+      <Locations />
       <Footer />
     </div>
   );
