+++ src/Components/footer.js
@@ -6,7 +6,7 @@ const Footer = () => {
             <div className="container mx-auto px-4">
                 <div className="flex flex-col md:flex-row justify-between items-center text-center md:text-left">
                     <p className="font-bold mb-2 md:mb-0 underline">
-                        COPYRIGHT © 2024 MH SUB I, LLC DBA OFFICITE.
+                        © 2024 NTIPL amay.g@netlabindia.com.
                     </p>
                     <div className="flex flex-col md:flex-row space-x-0 md:space-x-4 mt-2 md:mt-0">
                         <a href="/admin" className="hover:underline text-white font-bold mb-2 md:mb-0">Admin Log In</a>
